Declare the Background Workers to be of specific types
This fix was prompted by the following message in the server logs.

> LOG:  background worker "??H??" (PID 27511) exited with exit code 1

Postgres version 11 introduced the bgw_type field, and it is intended to
be used for process grouping in pg_stat_activity view, and messages such
as the above. Since the BDR code does not initialize the struct with
zeroed-out memory, the Postgres code uses the uninitialized struct
member's contents and prints garbage.

We initialize the struct with zeroes, so that any uniniitialized fields
introduced in future don't get populated with garbage. And then we
assign appropriate names and types to the workers.

Additionally, this commit removes some code and comments that are not
necessary anymore, as suggested by Bharath Rupireddy. E.g. remove all
traces of bgw_main, remove superfluous assertions, etc.