Setup CI pipeline for running BDR tests
This commit sets up CI pipeline for BDR - a github workflow to
run regression tests (both SQL tests and TAP tests).

It currently runs tests on upstream postgres REL_11_STABLE branch
and is likely to be extended for other branches as well once BDR
code is extended to work on those branches.

Tests are scheduled to run every 3 hours and upon a new push or
PR. The 3 hourely run might have to change, but for now, it helps
to catch any test failures.