Assorted improvements and fixes to TAP tests
This commit provides assorted improvements to TAP tests found
while debugging occasional/sporadic/intermittent failures in some
of the TAP tests. A few of the fixes help prevent some occasional
failures. However, the occasional failures aren't gone completely,
it's just that the frequency of failures is reduced with these
fixes.