Fix flaky test in part_bdr.sql
A SQL test for checking status of the parting node in
part_bdr.sql is flaky i.e. sporadically fails with
ERROR:  Timed out waiting for slot drop. There's a good reason
why this can happen and the same part node status check code in
TAP tests explains the reason clearly:

It is unsafe/incorrect to expect the parted node to know it's
parted and have a 'k' state. Sometimes it will, sometimes it
won't, it depends on a race between the parting node terminating
its connections and it receiving notification of its own parting.
That's a bit of a wart in BDR, but won't be fixed in 2.0 and is
actually very hard to truly "fix" in a distributed system. So we
allow the local node status to be 'k' or 'r'.

This commit changes parting node status check code in the SQL test
to be similar to TAP tests.