Adding PG 15 compatibility.
It does compile for PG 11 and PG 15.

Make check (that does include TAP tests) are all green for both versions.

The code has been updated so that a single branch can be used for both versions
(and for 12, 13, 14 and others that we'll support).

The TAP tests for PG 11 are treated as exceptions (have been moved to
compat/11/t).

This commit/PR adds some XXX to highlight places for review.