Allow memory allocations within a Critical Section
Typically, postgres doesn't allow pallocs in a CS for the fear
that any memory allocation failures can lead to PANICs. However,
in bdr_locks.c BDR wants to commit the txn in a CS, and
CommitTransactionCommand() does some pallocs causing assertion
failures in mcxt.c. Intention of the CS is that any errors must
lead to crash/restart to retry the txn.

Therefore, let's allow pallocs in CS for the current memory
context to preserve what's needed for the BDR in that particular
CS.