Fix pgbench with BDR testing script
There exists a shell script to run pgbench tests with BDR which is
invokable via make. However, in its current shape it won't work
just with make. Because, it requires both postgres and BDR source
code trees to exist, and it tries to get them from postgres
regression framework, which it can't because pg_config doesn't
have the source code tree information. Even if we make it work
somehow, which is complex IMO, the conf file used by script has
some removed parameters (such as bdr.connections,
bdr.<<node_name>>_dsn etc.). Moreover, this script is
developer-friendly, not for production uses, and has some nice
automation to save developers time, so NO to get rid of it
completely.

I'd like to propose it in a new form - a standalone script without
any dependency on make or core regression test infrastructure so
that developers can use it according to their tastes. FWIW, to
test for the new script's correctness a CI step is added.