Refactor various .conf files handling
BDR tests use custom configurations, most of which are spread
across many .conf files with a few interesting pieces. For
instance, there's a common .conf file for
track_commit_timestamp = on, wal_level = logical,
shared_preload_libraries = 'bdr', and is included in other conf
files. And, there's a .conf.sample to showcase the minimum
settings needed for BDR to work. When we have the docs covering
what's needed, no point of having a separate .conf.sample file.
And, there's a pg_hba.conf file which is exactly the same as that
postgres generates by default (perhaps, it was needed back then,
but no more it is required).

So, there's no point in maintaining these many .conf files in the
source code unnecessarily. This commit simplifies .conf file
handling one file per test, and gets rid of conf.sample and
pg_hba.conf. This makes the source code/tree look cleaner with a
bit less files (2).