Unregister BDR workers to prevent fail-and-restart cycles
After a BDR node is parted or BDR is removed from local node,
per-db and apply workers enter fail-and-restart cycles, IOW, the
workers will fail in any of the subsequent operations with either
ERROR or FATAL without unregistering themselves from BDR shared
memory causing the workers to be restarted by postmaster.
Typically, ERROR or FATAL in any bg worker will lead to
proc_exit(1) and, the postmaster restarts bg worker with
bgw_restart_time unless the bg worker exits with proc_exit(0).

For instance, if the per-db worker isn't unregistered it fails
with "ERROR:  could not find node (...)".

Similarly, if the apply worker isn't unregistered it fails with
"ERROR:  failed to find expected bdr.connections row ...".

A simple way to fix this is by checking node status in
bdr_bgworker_init() i.e. before the worker gets to real business
and, unregister it if node status is killed or no row exists for
the node in bdr.bdr_nodes.

A TAP test verifying this use-case is added.