Get rid of bdr_initial_load script to use system()
BDR uses bdr_initial_load script to execute pg_dump and pg_restore
while initializing a local node with remote node contents as part
of local node joining with bdr_group_join() i.e. logical join. The
per-db worker prepares necessary commands, forks a separate child
process to execute the bdr_initial_load script, and the per-db
worker waits for the command execution to finish. And, this part
of the code isn't coded with portability in mind (it won't work on
Windows).

What we need here can just be achieved with system() commands i.e.
pg_dump and pg_restore commands can just get executed by the
per-db worker itself with system() commands. This is similar to
what core postgres does with commands, for instance
archive_command, restore_command and so on.

With this commit, there is no need for the per-db worker to fork a
child process, no need to maintain separate scripts in the source
tree.