Fix an assertion failure while setting bdr.do_not_replicate
The Assert(!IsBackgroundWorker); in
bdr_do_not_replicate_check_hook fails when per-db is doing
pg_restore on the local node. Because, the pg_restore executes a
bunch of SQL statements from the pg_dump of the upstream node, and
the SQL statements can generate query plans with parallel workers.
Since every parallel worker loads GUC state
(ParallelWorkerMain->RestoreGUCState), and we have
bdr.do_not_replicate set to on during pg_dump/pg_restore, the
assertion fails.

Per commit  b21b2c8b203698d837daedff127007a9a550979b, intention of
the bdr.do_not_replicate GUC is to not replicate changes from
local node while it is initing from upstream node. The GUC is
supposed to be used internally, and not by the users.

The assertion failure is seen when pgbench is run on an upstream
node with '-i -s 100', and the local node is joined using
bdr_group_join() to the upstream node.

To fix the assertion failure,  let's get rid of
Assert(!IsBackgroundWorker);, and allow the GUC to be set only
when the local node is in bdr_init_exec_dump_restore().