add bdr.log_min_messages GUC
The BDR bgworker are initialized without a username in BackgroundWorkerInitializeConnection().

One consequence is that the log_min_messages level inherited by the bgworkers is coming from:

 - the superuser's rolconfig that set log_min_messages (if any) that created the extension
 - the log_min_messages GUC value if the superuser that created the extension does
   not have a rolconfig that set log_min_messages to another value

For example, if the superuser that created the extension executes:

alter user <my_superuser> set log_min_messages = PANIC;

then, after a restart, the bgworkers would not display any messages in the logfile
despite the log_min_messages set to WARNING in the postgresql.conf file.

This patch ensures that the new GUC "bdr.log_min_messages" will always been used
as the log_min_messages level for the BDR bgworkers (even if the superuser that
created the extension does have a rolconfig that set log_min_messages to another value).