Allow BDR to generate its own node identifier
BDR generates a new node id using bdr_init_copy, which is expected
to run on a node before creating a BDR group or joining the node in
an existing BDR group. bdr_init_copy overwrites the pg_control
file's system_identifier field. It uses a modified version of
core's pg_resetwal to overwrite the pg_control file.

There are some disadvantages with the above approach:
- Every node has to be BDR-ized with bdr_init_copy making a direct
  physical base backup of a node unusable as-is needing extra
  steps in between.
- pg_control file overriding is no good for production workloads,
  and may have unintended consequences - what if this is an
  already BDR active cluster being joined.
- Unnecessarily an out-of-core version of pg_resetwal needs to be
  maintained for BDR-izing.

To overcome the above issues, we make BDR generate its own node
identifier, and store it in a control file under node's data
directory. BDR generates a new node everytime when a BDR group is
first created on the node or the node is being added to an
existing BDR group for the first time. This generated node id
is used throughout the lifetime of the BDR on the node. The BDR
control file is removed when BDR is last removed on the node
(bdr.remove_bdr_from_local_node() is run for the last BDR-enabled
database on the node). If the BDR control file is removed in
between for any reason, BDR stops working.

We introduce SQL-callable functions to manage the BDR control
file - generate/create, read and remove the file. SQL-callable
functions are necessary so that the remote nodes, and other
components like bdr_init_copy can use it to get the node id and/or
manage it. For safety reasons, the execute permissions are
revoked from public, and if need be, appropriate roles can be
granted permissions.

Existing SQL/TAP tests cover the code added, hence no need for
extra tests.