Adressing new comments on the re-join behavior
There is no need to check for RecoveryInProgress() before dropping the replication
origins (as a similar check is already done in bdr_supervisor_worker_main()).

Renaming bdr_dont_insert_or_delete_killed to bdr_handle_rejoin.

The UPDATE added in bdr_handle_rejoin will be needed for the new "Allow BDR to
generate its own node identifier" #75 PR coming soon.

Once the PR #75 gets in, a re-join not using the previous node name will keep the
previous 'k' entry on the non parted nodes (while the re-joining one won't get it).

That should not be an issue at all, as 'k' entries are discarded in BDR anyway,
but we can re-visit this approach if the logical rejoin feature gets used heavily
and those remaining 'k' entries lead to confusion.