Upload only log files to artifactory in CI on failures
CI workflow currently uploads entire tmp_check directory upon
failures making the artifactory sizes to be more 1 GB. Most of it
are temporary pgdata directories created for each of the TAP test
which are unnecessary. What we are interested in for analysis of
any failures is the server logs and regression tests logs which
are avaiable under tmp_check/log. Therefore, upload only the logs
(a few MBs in size). This also reduces the upload artifactory step
time in CI.