Rename BDR SQL functions for better consistency
In this commit, some of the user-facing extension SQL functions
are renamed to be more redable and consistent. For instance, every
function is prefixed with 'bdr_',  and every internal function
(that's not supposed to be used outside by the end user) is
prefixed with '_', and suffixed with '_private'.

These function names have been agreed upon internally. And, the
original function names and their new names are documented for
reference with previous BDR versions.

Note that the code changes in this commit are majorly of type
s/old_function_name/new_function_name and hence are mechanical.