Change global_seq_nextval to bdr_snowflake_id_nextval
Essentially, global_seq_nextval that BDR exposes to the user
generates a snowflake ID with node timestamp, unique node sequence
ID and the nextval() of the local sequence. It's better to call
that it's a snowflake ID and follow the function naming
convention.

This commit also enables ddl_sequence.sql and
ddl_fn/ddl_sequence.sql test to be run, which are so far not
enabled to run.