Rename BDR SQL functions for better consistency (#117)
* Rename BDR SQL functions for better consistency

In this commit, some of the user-facing extension SQL functions
are renamed to be more redable and consistent. For instance, every
function is prefixed with 'bdr_',  and every internal function
(that's not supposed to be used outside by the end user) is
prefixed with '_', and suffixed with '_private'.

These function names have been agreed upon internally. And, the
original function names and their new names are documented for
reference with previous BDR versions.

Note that the code changes in this commit are majorly of type
s/old_function_name/new_function_name and hence are mechanical.

* Update a warning about usage of internal functions

Per review comment.

* Change global_seq_nextval to bdr_snowflake_id_nextval

Essentially, global_seq_nextval that BDR exposes to the user
generates a snowflake ID with node timestamp, unique node sequence
ID and the nextval() of the local sequence. It's better to call
that it's a snowflake ID and follow the function naming
convention.

This commit also enables ddl_sequence.sql and
ddl_fn/ddl_sequence.sql test to be run, which are so far not
enabled to run.

---------

Co-authored-by: Bharath Rupireddy <bharath.rupireddyforpostgres@gmail.com>