Provide a GUC to configure maximum allowed nodes in BDR
Invent bdr.max_nodes GUC to to configure maximum allowed nodes in
a BDR group. This guards against users adding unreasonably higher
nodes to a BDR group, which eventually will end them up in a
poorly maintained setup. A node fails to join a BDR group if
bdr.max_nodes has a different value for this parameter when
compared with its upstream node. Users must ensure all BDR members
have the same value for the parameter at any point of time,
otherwise, joining of nodes to the BDR group will start failing.
Default value for this parameter is 4, meaning, at any point of
time, there can be maximum of 4 nodes allowed in the BDR group.
Set a reasonable value to this parameter, only if required. Note
that more members in a BDR group require a more sophisticated
monitoring and maintenance, so choose this parameter value with
care.