Disallow BDR pulling in changes from non-BDR logical replication sources
We disallow a BDR node being a subscriber in postgres logical
replication or creating an external logical replication extension
(pglogical for instance) when BDR is active. Technically, BDR has
nothing to do with such external logical replication sources,
however, we disallow them for now to not have any possible data
divergence issues and conflicts on nodes within the BDR group. For
instance, when a BDR node pulls in changes from a non-BDR node
using any of external logical replication sources, then, the node
can easily diverge from the other nodes in BDR group, and may
cause conflicts.

However, we have no problem if a BDR node being a publisher in
postgres logical replication. Meaning, a non-BDR node can still
pull in changes from a BDR node. A test case to demonstrate this
use case is added.

That said, we might have to leave all of these to the user and
allow such extensions at some point.