Have a new initial extension version for BDR
Have a new initial extension version 2.1.0 for BDR because many
new features/bug fixes/improvements went into from the base repo
picked. The BDR extension versioning strategy is as follows:
- Version is of the 3 octet form.

- For a minor release i.e. any bug fixes, code or docs
improvements, performance work, the third octet will be
incremented - 2.1.1, 2.1.2 and so on.

- For a major release i.e. anything that affects compatibility,
nontrivial changes to extension setup script, new user-facing
function/GUC changes, the second octet will be
incremented - 2.2.0, 2.3.0 and so on.

- For a release that largely changes the BDR code i.e. revamping
BDR output plugin, a feature-complete DDL replication, automatic
global sequences, binary upgrade of BDR nodes etc., the first
octet will be incremented - 3.0.0, 4.0.0 and so on.

Note that to help with better reporting of issues prior to
external release, BDR_VERSION in bdr_version.h.in is suffixed with
"-RC1" which will be incremented "-RC2", "-RC3" for internal
releases. This prefix is to be removed before the external release
of the extension.