Fix pg_upgrade issue with a replication origin function (#151)
BDR defined a wrapper to get replication origin id of a given
transaction. The wrapper was defined in SQL when
pg_catalog.pg_xact_commit_timestamp_origin available (for
PG versions >= 14), otherwise defined in C by BDR (for
PG versions < 14). This was causing pg_restore to fail when a BDR
node is upgraded from say PG11 to PG15.

To fix the issue, we define the function to get replication origin
of a transaction in C within BDR for all of the versions, so while
upgrading the function exists in BDR shared library.

While here, the function is renamed to bdr_xact_replication_origin
to match with Postgres replication origin related functions.

Co-authored-by: Bharath Rupireddy <bharath.rupireddyforpostgres@gmail.com>