Avoid indefinite wait loops (#155)
* Avoid indefinite wait loops

It was observed that sporadically CI tests were running
indefinitely. So, looking at the possible indefinite wait loops,
two things came out:

1. A callsite of wait_for_pg_isready() in TAP tests isn't passing
a timeout which might cause an indefinite loop. So, changed
wait_for_pg_isready() to use TAP tests default timeout if no
timeout is passed.

2. The global DDL lock acquirer currently waits indefinitely for
global lock acquisition. This means that if the upstream node is
down or not responding for some reason, the acquirer waits
indefinitely. So, added an assert-only GUC for timing out for this
lock acquisition waiting. This is assert-only for now because the
CI tests use assert build, might need to turn it into a production
GUC if any issue is reported in future.

Note that these are two possible areas causing the TAP tests to
run indefinitely. These may or may not solve the problem, but they
have their own advantages of avoiding possible indefinite wait
loops.

* Reword comment

Per review comment.

---------

Co-authored-by: Bharath Rupireddy <bharath.rupireddyforpostgres@gmail.com>