Add pg16 support for bdr (#162)
* Add PG16 support for BDR

This commit adds Postgres version 16 support for BDR. As of this
commit, the Postgres REL_16_STABLE branch is on the commit
f75595cd [1] and PG_VERSION "16beta2".

[1]
commit f75595cd80d134d5a63929a90a0347ff9729b79a
Author: Tom Lane <tgl@sss.pgh.pa.us>
Date:   Fri Jul 21 12:00:14 2023 -0400

    Fix calculation of relid sets for partitionwise child joins.

* Increase too conservative timeouts for TAP tests

A few instances of timeouts for operations like concurrent
physical join, inserts, detach etc. are using too conservative
timeout 10 sec. This timeout isn't enough on slower machines, for
instance, on Ubuntu with 2 core, 8GB RAM. As a result, the TAP
tests will fail. This commit uses the postgres TAP test provided
default timeout (which is typically 180 sec) instead of mere 10
sec.

The tests failures were found and diagnosed on top of PG16 support
for BDR commit, hence the fix is also provided alongside the same
commit.

* Refresh pg_dump changes required for BDR

* Apply pg_dump changes required for BDR per commit 3c9965fe

* Add pg_dump base commit to README.pg_dump

* Remove extra empty line at the end of README.pg_dump

---------

Co-authored-by: Bharath Rupireddy <bharath.rupireddyforpostgres@gmail.com>