Add pgindent check to CI workflow
This commit runs pgindent on BDR code with PG16 branch where
postgres maintains pg_bsd_indent tool in code. If any pgindent
diffs are found, the CI run on PG16 fails, and the onus is on
developers to fix the indentation in their code. This saves others
time in running pgindent and keeps the coding standards high on
par with postgres. This is inspired by similar work postgres
did - where a separate buildfarm member was introduced to complain
if pgindent check fails.