Add pgindent check to CI workflow (#163)
* Add pgindent check to CI workflow

This commit runs pgindent on BDR code with PG16 branch where
postgres maintains pg_bsd_indent tool in code. If any pgindent
diffs are found, the CI run on PG16 fails, and the onus is on
developers to fix the indentation in their code. This saves others
time in running pgindent and keeps the coding standards high on
par with postgres. This is inspired by similar work postgres
did - where a separate buildfarm member was introduced to complain
if pgindent check fails.

* Make test -s return exit code 0 when diffs file doesn't exist

When pgindent.diffs file doesn't exist or exists without any
content, the test -s pgindent.diffs returns with exit code 1.
CI treats exit code 1 as a test failure and stops the entire
workflow. Hence, add exit 0 to fix this.

---------

Co-authored-by: Bharath Rupireddy <bharath.rupireddyforpostgres@gmail.com>