New approach to BDR node identifier generation (#168)
* New approach to BDR node identifier generation

BDR currently stores generated node identifier in a cluster level
BDR control file. Whoever needs it read from the control file. One
problem with the file-based approach is that the BDR control file
never gets replicated to streaming standbys connected to a BDR
node. Whereas, all the other BDR related tables, functions, in
fact extension itself gets replicated to standby via streaming
replication. The BDR control file needs to be manually copied to
standbys that are potential candidates for new primary after
failover. Otherwise, the new primary will have to discard the BDR
on it entirely (the data received so far into a BDR database also
needs to be dropped), and join freshly to the BDR group which can
be time-consuming and unnecessary.

This commit stores the BDR node identifier into a static getter
function in BDR schema so that the function gets replicated to
standbys via streaming replication. This getter function approach
avoids manual steps to ensure standbys get the same BDR node
identifier as that of the primary which is a BDR node. This paves
the way for supporting failover of BDR node in streaming
replication without needing to discard the entire BDR data, and
data in a BDR enabled database upon failover.

Essentially, the static getter function is per-database not the
per-cluster. Therefore, every database joining BDR group will get
a unique BDR node identifier unlike file-based approach to store
BDR node identifier. This makes dboid and timeline ID being used
to uniquely identify a BDR node (BDR-enabled database) irrelevant,
removal of which is parked for another day, not in this commit.

* Update docs related to BDR node identifier

* Add wait for standby catchup in newly added TAP test

---------

Co-authored-by: Bharath Rupireddy <bharath.rupireddyforpostgres@gmail.com>