Add support for seamless failover of bdr node to streaming standby (#169)
* Make TimeLineID use in BDR irrelevant

BDR uses postgres cluster's TimeLineID for the following
purposes:

- To uniquely identify a BDR node.

- To break ties during conflict resolution when local and remote
transactions timestamps are the same.

Using TLI makes it hard to support seamless failover of a BDR node
to streaming standbys. This is because the new primary gets a new
timeline after  failover which renders the BDR information
available on the new primary useless unless updated manually.

Therefore, this commit makes TLI irrelevant for BDR by using a
fixed BDR-level TLI value of 1. This is to be less intrusive as
far as TLI related code in BDR is concerned. With the TLI code in
BDR spread across most of the places, removing it entirely makes
it a huge change. Hence, a fixed BDR-level TLI value is used to
keep things simple. However, the TLI code may have to be fully
removed in future.

With the introduction of per-database unique BDR node identifier
by the commit d2945407, TLI component is no longer needed to
uniquely identify a BDR node.

To break ties during conflict resolution, we rely on a uniquely
assigned monotonically increasing node_seq_id value. BDR assigns
node_seq_id value for each BDR node. To keep things simple and
consistent across all BDR nodes, we break the ties with the
highest node_seq_id wins approach.

* Simulate seamless BDR node failover to streaming standby

This commit adds a TAP test to simulate seamless BDR node failover
to new primary, that is, promoted streaming standby after
failover.

* Make node with lowest node_seq_id to win conflict

Change the conflict resolution so that the node with lowest
node_seq_id wins. The idea is to give priority to an older BDR
node.

* Restore ThisTimeLineID assert checks

Previous commit changed a few asserts whose actual intention were
to check the postgres cluster TLI > 0, to BDR TLI > 0. Let's
restore such asserts.

* Use 0 value for BDRThisTimeLineID

Since TimeLineID is irreleveant for BDR, use an invalid value 0
for BDRThisTimeLineID.

---------

Co-authored-by: Bharath Rupireddy <bharath.rupireddyforpostgres@gmail.com>