Ensure skip_dll_replication has same value on all nodes
A new node fails to join a BDR group if it has a different value
for this parameter when compared with its upstream node. An existing node
can't start BDR workers if the parameter value doesn't match with its
upstream node. Hence, users must ensure all BDR members have the same value
for the parameter at any point of time.