Improve GUCs code
This commit improves the code around BDR GUCs:
- Renames some of debug-only GUCs to debug_<<name>> for better
readability and reflection of why the GUC is for.

- Makes GUCs description provided in DefineCustomXXX consistent.

- Improves some of GUCs description provided in DefineCustomXXX.

- Removes "bdr:" from some of the error messages to be consistent
with elsewhere.

- Removes log_min_messages = debug2 from TAP tests to make things
faster by getting away with TAP tests logging a lot of info for
every node started by TAP framework.

- Makes bdr.permit_node_identifier_getter_function_creation more
strict for end-users by adding show to superusers only flag, and
disallowing one to set it in conf files as the GUC is not indended
for end-users.