Individualize temp dump directory by using system identifier
In CI runs of TAP tests, it was observed that a supervisor worker
in one postgres cluster is cleaning up temporary dump directories
of other postgres clusters. This is because the clean up logic
assumes a fixed prefix for all temp dumps across all postgres
clusters being run on the same host system. Therefore,
individualizing temp dump directory with postgres
system_identifier that's unique per postgres cluster is chosen as
the way to fix this issue.