Fail fast when non-empty database tries to join BDR group
This commit adds a check to see if the database that's joining BDR
group has any existing tables so that the join can fail fast.
When a node joins BDR group with pre-existing tables, issues will
start to appear at a later point of time making it hard to detect
and reason about.

A TAP test is added to verify the failure. For this, an existing
TAP test is renamed to be more generic.