Address review comments
Check for free space on data directory file system, if free space
is 1.2X the remote database size, emit a warning. 1.5X seemed too
large in production systems (I can't imagine postgres needing
500GB for temp work space for indexes, storing intermediate query
results etc. for a remote database of size 1TB).

Check for free space on temp dump directory file system, if free
space is less than 50% of remote database size excluding the
total indexes size, emit a warning.

Check if data directory and temp dump directory are on same file
system mount point, if yes, check if free space on the file system
is not less than 1.2X the remote database size and remote database
size excluding the total indexes size (combination of above two
checks basically as the same file system has to accommodate both
temp dump and database).

Note that everything is a WARNING not an ERROR because we are
dealing with estimations, so we can't be so perfect but we can
always inform early.