Fail fast with non empty database joining bdr group (#186)
* Fail fast when non-empty database tries to join BDR group

This commit adds a check to see if the database that's joining BDR
group has any existing tables so that the join can fail fast.
When a node joins BDR group with pre-existing tables, issues will
start to appear at a later point of time making it hard to detect
and reason about.

A TAP test is added to verify the failure. For this, an existing
TAP test is renamed to be more generic.

* Add bypass user tables check option for bdr_init_copy

Database joining BDR group physically will essentially have
existing user tables on the database. So, add a bypass option to
bdr.bdr_join_group to skip user tables check and pass it as true
for physical joins in bdr_init_copy.

* Modify the failing TAP tests with user-table precheck

* Detangle user tables check from remote_dsn

* Address review comments

- Reword message text
- Correctly create 2 nodes in tests

---------

Co-authored-by: Bharath Rupireddy <bharath.rupireddyforpostgres@gmail.com>