Fix sporadic test failure in 045_unregister_workers_after_detach.pl
Tests in 045_unregister_workers_after_detach.pl sporadically fails
to detech the unregistering worker message in server logs.
Sometimes it may happen that the worker might get killed even
before unregistering log message is hit. So, skip the check for
unregistering worker message when it is not found in server log.