Add code coverage GitHub CI action (#187)
This commit adds a GitHub CI action for running code coverage on
BDR source code. It produces the code coverage final html files as
downloadable artifacts which can be viewed in web browser locally.
One can run the code coverage manually by going to GitHub portal's
Actions tab and selecting BDR code coverate CI workflow on the
left side of the web page, and clicking run workflow on the right
side of the web page.

Running BDR tests with --enable-coverage unveiled some test
failures that can occur if starting up of BDR workers is slow. To
fix those failures, this commit increases the sleep time in tests.
However, the failures might have to be fixed in a different way in
future, but that's for another day.

Co-authored-by: Bharath Rupireddy <bharath.rupireddyforpostgres@gmail.com>