Allow specifying user mapping alongwith foreign server in DSNs
Commit ffea0747f4 introduced support for user mapping and foreign
server in place of DSNs. With that commit, the foreign server name
was passed in and current user id is obtained in to get the
associated user mapping. This is a problem because a single
foreign server can be associated with multiple user mappings and
all the pgactive bg workers return the user that they are
connected to the database, but not the actual user running the
create group or join group query. Also, the full connection string
is derived from both user mapping and foreign server. Therefore,
this commit supports specifying both user mapping and foreign
server as a key-value pair which then be used to derive connection
string without relying on the current user of the pgactive bg
workers.

In passing, also remove an assertion which was failing when a
SQL function is dropped.

Bump pgactive_VERSION to 2.1.0-RC3.