Support for seamless pg_upgrade of pgactive nodes
This commit does the following things to support seamless
pg_upgrade of pgactive nodes. Onus is on pg_upgrade to verify and
copy the pgactive replication slots to new cluster/upgraded database
instance from old cluster.

- When pg_upgrade creates pgactive logical replication slots on
the new cluster, it tries to set up the decoding context and checks
if the pgactive output plugin looks okay. When doing so, it ends up
calling pgactive output plugin callback pg_decode_startup where the
pgactive shared memory is accessed which isn't yet fully set up.
Therefore, this commit skips pgactive shared memory access in
pg_decode_startup and pg_decode_shutdown when in binary upgrade
mode. Later when the pgactive replication is actually started these
callbacks are anyways called to set things up on pgactive shared
memory.

- When the new cluster is up and making pgactive replication
connection to remote node to create replication slot, it fails with
slot already exists error. This is because the replication slot
pertaining to the new cluster on the remote node already exists.
This commit drops the slotif it exists and recreates it.

To illustrate, consider a pgactive group with 3 nodes n1, n2 and
n3. The node n1 will have slots for n2 and n3. The node n2 will
have slots for n1 and n3. The node n3 will have slots for n1 and
n2. Now, upgrade node n2, call it n2new and the upgrade will have
the slots for n1 and n3 restored so that n1 and n3 will be able to
get the changes seamlessly. But, when the n2new is up and try to
set up replication origins for remote nodes n1 and n3, make
replication connections to remote nodes n1 and n3, but it fails
with "replication slot already exists". Note that the upgrade
process is nothing to do with n2 replication slots on n1 and n3.
And, pgactive up until this commit doesn't support drop and
recreate slots. With this commit, the pgactive drops and recreates
slots if they exist on respective remote nodes.

- This commit requires pgactive to fail fast if a database that's
joining pgctive group is already part of another pgactive group.
The easiest way is to check if pgactive replication slot already
exists for this database. Before this commit, the failure occurs
much later i.e. when the replication slot for this node is created
by upstream node.