Fix corner case when checking params are same on all nodes
The corner case depends in which node order the check is done when comparing
parameters values.

Example:

Node 1, 2 and 3 have param par set to val1.

Node 2 changes param par to val2 -> compares with Node 1 => not ok
Node 1 changes param par to val2 -> compares with Node 2 => ok
Node 2 re-check with Node 1 => ok

Now, Node 1 and Node 2 have val2 and Node 3 has val1 (which is not what we want).

This commit ensures that we are comparing with all the nodes (before we were fine
if we found one node that had the same value as our node).

It does fix https://github.com/aws/pgactive/issues/219