Add all postgres versions support for pgactive (#214)
* Add PG13 support for pgactive

This commit adds Postgres version 13 support for pgactive. As of
this commit, the Postgres REL_13_STABLE branch is on the commit
4caa9e33ecfe70359177b9689f33c6da55c3aa35 and PG_VERSION "13.12".

* Add PG14 support for pgactive

This commit adds Postgres version 14 support for pgactive. As of
this commit, the Postgres REL_14_STABLE branch is on the commit
5a32af3f2c6106080da72e78a42c8fc3ab0584a1 and PG_VERSION "14.9".

* Add PG12 support for pgactive

This commit adds Postgres version 12 support for pgactive. As of
this commit, the Postgres REL_12_STABLE branch is on the commit
01bc42634e59e1e746eec236ac17247bdcc228e9 and PG_VERSION "12.16".

* Remove unneeded definition of AT_ProcessedConstraint

AT_ProcessedConstraint was removed in PG13 by
1281a5c907b41e992a66deb13c3aa61888a62268, so no necessary to have
its traces in versions other than PG11 and PG12.

* Add PG17 (17devel) support for pgactive

This commit adds Postgres version 17 support for pgactive. As of
this commit, the Postgres HEAD/master branch is on the commit
a0a5e0feb35c1209e340046dc4e40af3b0e79d59 and PG_VERSION "17devel".