Remove unneeded pgactive_compat.c (#221)
This commit removes unneeded pgactive_compat.c from all
src/compat/<PG VERSION> directories because the actual
compatibility macros or inline functions are in pgactive_compat.h.

It also removes pgactive_compat.c from src directory and moves
that code to pgactive.c.

It also fixes an issue with pgactive_get_integer_timestamps which
is using a macro that doesn't exists in PG or in pgactive. PG
seems to be using a GUC integer_datetimes to determine if the
server installation uses integer datetimes or not. The function
pgactive_get_integer_timestamps is now made to read the PG GUC
integer_datetimes.