Handle the new "flush" parameter in LogLogicalMessage
Core PostgreSQL added a new "flush" parameter to LogLogicalMessage
in 173b56f1ef.

Basically it requests to flush the lsn in case this new parameter is set to
true and the existing "transactional" parameter is set to false.

We are already requesting the lsn flush in pgactive_send_message(), so using
LogLogicalMessage() with flush being "false" looks fine.