Fix a few things with pgactive_fdw check added by commit 2a8474d7
Commit 2a8474d7 added a check to verify user-entered foreign
server is associated with pgactive_fdw, if not the create or
join group fails with an error with such FDW specified. However,
there's no test case added to verify the fix and cover the code
added. Also, it caused a memory leak for the cmd varialbe as it
forgot to convert initStringInfo(&cmd); for the pg_user_mappings
query to resetStringInfo(&cmd);.

This commit adds a test case to hit the error added by 2a8474d7
and fixes the memory leak issue.

In passing, it rewords the error message a bit to be consistent
with the PostgreSQL messaging style. Alos, it changes the code to
use ERROR instead of FATAL, because this code can be hit by a
direct user SQL, a FATAL emits
"server closed the connection unexpectedly" to the client
connected to the backend which is a common error even for SEGVs or
any other errors.