Simplify parameter check from other nodes code
This commit simplifies code in check_params_are_same() using list
data structure APIs that postgres provides. Existing code is
hard to understand which uses boolean flags and a while loop and
a for loop. The core logic (that is to check parameters with all
of the connectable nodes and fail if none of the remote nodes is
connectable) of the function is changed to the following:
- Capture the length of the prepared remote nodes list.
- Iterate over the list to remove all the connectable nodes from
the list.
- If none of the remote nodes is connectable in which case the
prepared list length remains the same, fail with a FATAL.