Improve commit txn code while removing pgactive_locks entry
To release global lock, pgactive node ensures its entry is removed
from pgactive_locks before releasing the in-memory lock; if any
error occurs while committing the txn, it wants to crash the
server and restart the event again. To achieve the crash and
restart, the txn is committed inside a critical section
(note that ERROR/FATAL is converted to PANIC in a CS). Since
interrupts are blocked in a CS and given the fact that the commit
txn can do a bunch of other things that may require interrupts
to be processed, committing within a CS is not a good way.

This commit moves the commit txn to PG_ENSURE_ERROR_CLEANUP and
PG_END_ENSURE_ERROR_CLEANUP blocks from CS to convert any ERROR or
FATAL to PANIC.