Add support for data-only logical join
This commit adds a GUC pgactive.init_node_data_only when set to
true the logical join syncs only the data from the upstream node
not the schema. User must ensure the joining node has all required
schema created before logically joining it to pgactive group.

Data-only logical join helps in situations where upstream node has
some of the database objects (functions, procedures etc.) that
can't be or not allowed to be synced to the joining node due to
security or various other reasons.

Following are the steps to do data-only logical join:
- Set pgactive.init_node_data_only to true on the joining node.
- Create all the required schema (tables definitions) on the
  joining node.
- Logically join the node to an upstream node.

Extension version is bumped to 2.1.4 as the existing user table
check in _pgactive_begin_join_private need to be skipped for
data-only logical joining.