Rethink the fix for deadlock in commit 97900b23
This commit rethinks the fix provided for deadlock reported in
commit 97900b23 in the context that it used LWLockHeldByMe for
pgactive_locks shared memory. The fact is that LWLockHeldByMe can
get costiler as it just loops over all the locks held by the
backend and it is supposed to be used only for debug purposes.

To avoid the deadlock reported in commit 97900b23, alternate
simple fix is to just not use pgactive_NODEID_FORMAT_WITHNAME_ARGS
to report node name in pgactive_locks_node_detached, but using
node id there is sufficient in the log messages.

While on this, remove another usage of LWLockHeldByMe in
pgactive_get_last_applied_xact_info for the same reason that it is
supposed to be used only for debug purposes.