Fix supervisor issues in per-db worker registration failure
This commit fixes a few issues in the way supervisor reacts when
it can not register per-db worker due to background worker
exhaustion in postgres. Currently, the supervisor fails with an
ERROR in this situation without freeing up the per-db slot from
pgactive shared memory. This causes exhaustion of all the worker
slots in pgactive shared memory. Moreover, emitting ERROR upon
background worker exhaustion in postgres causes the supervisor to
enter a restart loop which is not good.

This commit does the following in supervisor upon bakground worker
exhaustion in postgres:
- Free up the per-db worker slot in pgactive shared memory
- Emit WARNING to avoid restart loop

While here, this commit also does the following:

- Rewords certain error messages with a lookup table
- Changes per-db worker registration message levels from DEBUG to
  LOG to be more responsive and informative.